requirements.txt
matplotlib
missingno
nbformat
numpy
openpyxl
pandas
pmdarima
plotly
pyaml
Pillow
seaborn
statsmodels
scikit-learn
tqdm

absl-py==1.4.0
astunparse==1.6.3    
async-generator==1.10
attrs==23.1.0        
cachetools==5.3.0    
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0  
distlib==0.3.6
exceptiongroup==1.1.1      
filelock==3.12.0
flatbuffers==23.5.9        
gast==0.4.0
google-auth==2.18.0        
google-auth-oauthlib==1.0.0
google-pasta==0.2.0        
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
jax==0.4.9
keras==2.12.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.1
platformdirs==3.5.1
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytrends==4.9.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
termcolor==2.3.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
regex
